chore(tests): reorganize live test suite
