+++ tests/live/test_sdk_utilities_live.py
@@ -12,7 +12,7 @@ RUN_E2E = os.getenv("IMEDNET_RUN_E2E") == "1"
 pytestmark = pytest.mark.skipif(
     not RUN_E2E or not (API_KEY and SECURITY_KEY),
     reason=(
-        "Set IMEDNET_RUN_E2E=1 and provide IMEDNET_API_KEY/IMEDNET_SECURITY_KEY to run e2e tests"
+        "Set IMEDNET_RUN_E2E=1 and provide IMEDNET_API_KEY/IMEDNET_SECURITY_KEY to run live tests"
     ),
 )
 
@@ -27,7 +27,7 @@ def sdk() -> Iterator[ImednetSDK]:
 def study_key(sdk: ImednetSDK) -> str:
     studies = sdk.get_studies()
     if not studies:
-        pytest.skip("No studies available for e2e tests")
+        pytest.skip("No studies available for live tests")
     return studies[0].study_key
